+++ src/main/java/com/bulain/activiti/service/ReferanceServiceImpl.java
@@ -27,7 +27,7 @@ public class ReferanceServiceImpl extends PagedServiceImpl<Referance, ReferanceS
         return referanceMapper;
     }
 
-    public Referance get(Integer id) {
+    public Referance get(Long id) {
         Object object = cacheService.get(Referance.class, id);
         if (object == null) {
             object = super.get(id);
@@ -77,7 +77,7 @@ public class ReferanceServiceImpl extends PagedServiceImpl<Referance, ReferanceS
                 referance.getCategory());
         cacheService.delete(Item.class, key);
     }
-    public void delete(Integer id) {
+    public void delete(Long id) {
         Referance referance = super.get(id);
         super.delete(id);
