+++ src/test/java/com/bulain/activiti/dao/AccountMapperTest.java
@@ -20,7 +20,7 @@ public class AccountMapperTest extends ServiceTestCase {
 
     @Test
     public void testDeleteByPrimaryKey() {
-        int deleteByPrimaryKey = accountMapper.deleteByPrimaryKey(Integer.valueOf(101));
+        int deleteByPrimaryKey = accountMapper.deleteByPrimaryKey(Long.valueOf(101));
         assertEquals(1, deleteByPrimaryKey);
     }
 
@@ -44,7 +44,7 @@ public class AccountMapperTest extends ServiceTestCase {
 
     @Test
     public void testSelectByPrimaryKey() {
-        Account selectByPrimaryKey = accountMapper.selectByPrimaryKey(Integer.valueOf(102));
+        Account selectByPrimaryKey = accountMapper.selectByPrimaryKey(Long.valueOf(102));
         assertNotNull(selectByPrimaryKey);
 
         assertEquals("name_102", selectByPrimaryKey.getName());
@@ -54,7 +54,7 @@ public class AccountMapperTest extends ServiceTestCase {
     @Test
     public void testUpdateByPrimaryKeySelective() {
         Account record = new Account();
-        record.setId(Integer.valueOf(103));
+        record.setId(Long.valueOf(103));
         record.setName("name-updated");
         record.setType("type-updated");
         int updateByPrimaryKeySelective = accountMapper.updateByPrimaryKeySelective(record);
@@ -64,7 +64,7 @@ public class AccountMapperTest extends ServiceTestCase {
     @Test
     public void testUpdateByPrimaryKey() {
         Account record = new Account();
-        record.setId(Integer.valueOf(104));
+        record.setId(Long.valueOf(104));
         record.setName("name-updated");
         record.setType("type-updated");
         int updateByPrimaryKey = accountMapper.updateByPrimaryKey(record);
@@ -82,7 +82,7 @@ public class AccountMapperTest extends ServiceTestCase {
     @Test
     public void testCountByUniqueKey4Update(){
         Account record = new Account();
-        record.setId(Integer.valueOf(104));
+        record.setId(Long.valueOf(104));
         record.setName("name_103");
         Long cnt = accountMapper.countByDuplicate(record);
         assertEquals(Long.valueOf(1), cnt);
